// frontend/models/Blog.ts
export interface Blog {
	id: number;
	category: string;
	claps: number;
	content: string;
	dateCreated: string;
	displayChart: string;
	excerpt: string;
	featured: boolean;
	headerImg: string;
	icon: string;
	marketPair: string;
	popular: boolean;
	projectName: string;
	readTime: string;
	slug: string;
	summary: string;
	thumbnail: string;
	thumbnailAlt: string;
	ticker: string;
	title: string;
}